requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.7.0
Flask==3.0.3
frozendict==2.4.6
html5lib==1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
plotly==6.0.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
retrying==1.3.4
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.0.6
yfinance==0.2.52
zipp==3.21.0
